Add files via upload
In the final section of the book methods have depreciated and can no longer be used. The book was published in 2016 and the Text method was depreciated in 2019. This is the documentation I am going to include in the book for future readers potentially to have an easier time than I.